Allow TEXTAREAs to be updated (#10)
This fixes a bug noticed while building a demo.

It seems that all TEXTAREAs are considered deeply equal (isEqualNode) by the
DOM regardless of their value. This is very strange because the same thing is
not true for input or select. In fact, it is so strange that it is easier for
me to believe that there is something wrong with my understanding of the
situation (or perhaps a bug in morphdom) than this is everything working as
intended.

Regardless, adding this special case remedies the immediate issue.

If more of these special cases arise, it might be worth taking the performance
hit and using morphdom's native diffing which does also fix the issue.