Add support for accessing the currentTarget (#16)
This adds the ability to access the currentTarget of events. Since, in Motion, this is
the element to which the developer added the data-motion attribute, it is aliased to
element to make it clear that it is usually the correct "default" choice.

This implementation will probably include two copies of the same form data. I don't
think this is a huge concern though.

Co-authored-by: Drew Ulmer <drew@unabridgedsoftware.com>