Use faster but correct implementation for render_hash (#18)
This resolves a "TODO" left in code. I wasn't sure what the implementation of render_hash should be, but after some thought and experimentation, I feel good about this.

The previous implementation was "trivially correct but wasteful". In addition to using a cryptographic hash (which we do not need in this instance), it also encrypted and signed the state of the object just to throw it away.

As it turns out, Ruby's built-in hash is probably good enough for our purposes. It has 64 bits of entropy. This means the chance that any two distinct values (we only care about pairs) share a hash is 1:2^64 (which is very unlikely!).

I was also very pleasantly surprised to find out that what I considered to be the safer implementation is faster.