Add Callbacks API (`bind`) (#32)
Before this change, the only way for a child component to communicate with a
parent was with broadcasts. This works well when the purpose of communication
is to signal that an external resource (like a model) has been updated, but it
gets a bit awkward when the goal is to communicate something specific about the
UI state of the page (like a section being collapsed). In the latter case, you
need to generate and pass around a unique topic name for the broadcasts to
ensure any changes are scoped to that user's view.

This introduces a new Callbacks API that makes this less cumbersome.

Parent components can expose a method as a callback to their children using
`bind`, which be called both in the component class and the template:

```
class CounterComponent < ViewComponent::Base
  include Motion::Component

  attr_reader :count

  def initialize(count: 0)
    @count = count
  end

  def increment
    @count += 1
  end
end
```

```
<div>
  <h1>The count is <%= count %>!</h1>

  <%= render ButtonComponent.new(text: "+", on_click: bind(:increment)) %>
</div>
```

Child components can receive the callback created by bind just like any other
state, and trigger it using the `call` method.

* Support setting up new motions, broadcasts, and timers during render

Co-authored-by: Gwendolyn Holley (Wendy) <Gmfholley@gmail.com>