Don't use a layout when rendering for cable
In Rails < 6.1, ViewComponent includes a monkey patch for the
`render_in` method that bypasses the Rails rendered.

Unfortunately, the offical `render_in` method renders the component
within the context of the full page, rather than in isolation. When
morphdom receives a payload that is of a different type to the original
(full document vs element/node) it deletes the element on the page and
attempts (and fails) to replace it with the full document.

Using an anonymous class with layout set to false tells Rails not to
render the rest of the page